refactor: add mutation reveal votes

+++ internal/models/room_model.go
@@ -5,5 +5,5 @@ type Room struct {
 	Title       string
 	HostName    *User
 	Users       []*User
-	AverageVote float64
+	AverageVote float32
 }
