+++ internal/models/user_model.go
@@ -3,6 +3,6 @@ package models
 type User struct {
 	ID       int
 	Username string
-	Vote     int
+	Vote     float32
 	IsAdmin  bool
 }
